Make stuff compile, add first unit tests (WIP), some moving around
