+++ distribution/lib/Standard/Table/0.0.0-dev/src/Data/Serialization_Options.enso
@@ -1,3 +1,5 @@
+from Standard.Base import all
+
 ## Specifies options for reading text data in a table to more specific types and
    serializing them back.
