+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/AdditionalInvalidRows.java
@@ -1,4 +1,4 @@
-package org.enso.table.read.parsing.problems;
+package org.enso.table.parsing.problems;
 
 /** A problem which indicates how many additional invalid rows were encountered. */
 public class AdditionalInvalidRows implements ParsingProblem {
