+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/InvalidRow.java
@@ -1,4 +1,4 @@
-package org.enso.table.read.parsing.problems;
+package org.enso.table.parsing.problems;
 
 /** A problem indicating that a row contained more or less columns than expected. */
 public record InvalidRow(long source_row, Long table_index, String[] row) implements ParsingProblem {}
