+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/MismatchedQuote.java
@@ -1,4 +1,4 @@
-package org.enso.table.read.parsing.problems;
+package org.enso.table.parsing.problems;
 
 /** A problem indicating that a quote has been opened and never closed. */
 public class MismatchedQuote implements ParsingProblem {}
