+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/ParsingProblem.java
@@ -1,4 +1,4 @@
-package org.enso.table.read.parsing.problems;
+package org.enso.table.parsing.problems;
 
 /**
  * A parent class for parsing problems which may be reported as warnings or errors, depending on the
