+++ std-bits/table/src/main/java/org/enso/table/read/DelimitedReader.java
@@ -14,10 +14,11 @@ import org.enso.table.data.column.storage.Storage;
 import org.enso.table.data.index.DefaultIndex;
 import org.enso.table.data.table.Column;
 import org.enso.table.data.table.Table;
-import org.enso.table.read.parsing.problems.AdditionalInvalidRows;
-import org.enso.table.read.parsing.problems.InvalidRow;
-import org.enso.table.read.parsing.problems.MismatchedQuote;
-import org.enso.table.read.parsing.problems.ParsingProblem;
+import org.enso.table.parsing.ParsingFailedException;
+import org.enso.table.parsing.problems.AdditionalInvalidRows;
+import org.enso.table.parsing.problems.InvalidRow;
+import org.enso.table.parsing.problems.MismatchedQuote;
+import org.enso.table.parsing.problems.ParsingProblem;
 import org.enso.table.util.NameDeduplicator;
 
 /** A helper for reading delimited (CSV-like) files. */
