+++ std-bits/table/src/main/java/org/enso/table/read/WithProblems.java
@@ -1,6 +1,6 @@
 package org.enso.table.read;
 
 import java.util.List;
-import org.enso.table.read.parsing.problems.ParsingProblem;
+import org.enso.table.parsing.problems.ParsingProblem;
 
 public record WithProblems<T>(T value, List<ParsingProblem> problems) {}
